Move executive tests to own file
